pyproject.toml
[project]
name = "fft-viz"
version = "0.3.0"
description = "High-quality FFT visualization tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nfdpy",
    "pyviewer-extended",
    "requests>=2.32.3",
    "research-utilities",
    "setuptools>=72.1.0",
    "torchvision>=0.22.0",
]

[tool.uv.sources]
research-utilities = { git = "https://github.com/ShineiArakawa/research-utilities.git", tag = "v1.1.13" }
nfdpy = { git = "https://github.com/ShineiArakawa/nfdpy.git" }
pyviewer-extended = { git = "https://github.com/ShineiArakawa/pyviewer-extended.git", tag = "v0.1.3" }
